move .env to openbb_core root directory
