+++ openbb_sdk/sdk/core/openbb_core/app/model/charts/charting_settings.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 from typing import Optional
 
-from openbb_core.app.env import Env
 from openbb_core.app.logs.utils.utils import get_app_id
 from openbb_core.app.model.system_settings import SystemSettings
 from openbb_core.app.model.user_settings import UserSettings
+from openbb_core.env import Env
 
 
 # pylint: disable=too-many-instance-attributes
