+++ openbb_sdk/sdk/core/openbb_core/app/router.py
@@ -22,7 +22,6 @@ from pydantic.config import BaseConfig
 from pydantic.validators import find_validators
 from typing_extensions import Annotated, ParamSpec, _AnnotatedAlias
 
-from openbb_core.app.env import Env
 from openbb_core.app.model.abstract.warning import OpenBBWarning
 from openbb_core.app.model.command_context import CommandContext
 from openbb_core.app.model.obbject import OBBject
@@ -32,6 +31,7 @@ from openbb_core.app.provider_interface import (
     ProviderInterface,
     StandardParams,
 )
+from openbb_core.env import Env
 
 P = ParamSpec("P")
